refactor - refactor the project structure.
Make folders for pages and keep their relevant components in the folder rather than keeping it in the component file.

Add css file to each component.
