+++ src/App.css
@@ -4,23 +4,11 @@
 	--max-width: 1000px;
 }
 
-html {
-	scroll-behavior: smooth;
-}
-
-body {
-	min-height: 100vh;
-}
-
 .page-container {
 	max-width: var(--max-width);
 	margin: 0 auto;
 	padding: 1rem;
 	margin-top: 1rem;
-}
-
-.route-btn-group > button {
-	margin-right: 1rem;
 }
 
 .cards-display {
@@ -30,18 +18,6 @@ body {
 	margin-top: 1rem;
 	justify-content: center;
 }
-
-.cards-display-cart {
-	display: grid;
-	grid-template-columns: 1fr 1fr;
-	grid-gap: 2rem;
-	justify-content: center;
-	align-items: stretch;
-}
-
-/* .cards-display > div {
-	margin: 1rem;
-} */
 
 .out-of-stock {
 	position: absolute;
@@ -58,39 +34,6 @@ body {
 	z-index: 1;
 }
 
-.quantity-control {
-	border: 1px solid var(--primary);
-	align-self: baseline;
-	margin-left: 1rem;
-}
-
-.quantity-control > button {
-	background-color: var(--primary);
-	border: none;
-	height: 1.5rem;
-	width: 1.5rem;
-}
-
-.quantity-control > span {
-	padding: 0.18rem 0.25rem;
-}
-
-.header-links {
-	margin-left: auto;
-}
-
-.header-links > button {
-	background: none;
-	border: none;
-	margin-left: 1.3rem;
-	outline: none;
-	position: relative;
-}
-
-.header-links > button:hover {
-	color: var(--primary);
-}
-
 .icon-badge {
 	position: absolute;
 	display: flex;
@@ -105,24 +48,8 @@ body {
 	color: #fff;
 }
 
-.refine-search {
-	display: flex;
-	flex-wrap: wrap;
-}
-
-.filter-group {
-	display: flex;
-	flex-direction: column;
-	margin-right: 2rem;
-	/* border: 1px solid red; */
-}
-
 .card {
 	max-width: 350px;
-}
-
-.filter-group > p {
-	margin: 0.1rem 0rem;
 }
 
 .border-shadow {
@@ -133,29 +60,9 @@ body {
 	padding: 0.5rem;
 }
 
-.filter-group > label {
-	margin: 0.5rem 0rem;
-}
-
-.filter-group > label > input {
-	margin-right: 0.5rem;
-}
-
 .flex-space-between {
 	display: flex;
 	justify-content: space-between;
-}
-
-.show-search-options {
-	border: none;
-	background: none;
-	padding: 0.2rem;
-	display: none;
-}
-
-.cart-total {
-	display: flex;
-	flex-direction: column;
 }
 
 .margin-bottom-1 {
@@ -174,13 +81,6 @@ hr {
 
 .hr-dashed {
 	border-top: 1px dashed;
-}
-
-.cart-grid {
-	display: grid;
-	grid-template-columns: 2fr 1fr;
-	grid-gap: 1rem;
-	align-items: flex-start;
 }
 
 .image-container {
@@ -220,113 +120,10 @@ hr {
 	z-index: 2;
 }
 
-.header {
-	padding: 0rem 1rem;
-	flex-wrap: wrap;
-}
-
-.header > h1 {
-	font-family: 'Roboto', sans-seridf;
-}
-
-.header-container {
-	position: relative;
-	box-shadow: none;
-}
-
-.header-search-bar {
-	padding: 0.5rem 1rem;
-	margin: 0.5rem auto;
-	min-width: 300px;
-	display: flex;
-}
-
-.header-search-bar > input {
-	outline: none;
-	border: none;
-	margin-right: 0.5rem;
-	width: 100%;
-}
-
-.header-search-bar > button {
-	outline: none;
-	border: none;
-	background: none;
-	font-size: 1rem;
-}
-
-.header-search-bar > button:hover {
-	color: var(--primary);
-}
-
-.links-bar {
-	display: flex;
-	padding: 1.2rem 0rem;
-	position: -webkit-sticky;
-	position: sticky;
-	top: 0;
-	background-color: var(--primary);
-	max-width: var(--max-width);
-	margin: 0 auto;
-	z-index: 3;
-}
-
-.links-bar > button {
-	background: none;
-	border: none;
-	padding: 0rem 1rem;
-	outline: none;
-	position: relative;
-	font-size: 1rem;
-}
-
-.links-bar > button:hover {
-	font-weight: bolder;
-}
-
-.page-empty {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	align-self: stretch;
-	color: var(--gray-light);
-	min-height: 50vh;
-}
-
-.page-empty > h3 {
-	color: inherit;
-}
-
 @media only screen and (max-width: 500px) {
 	.cards-display {
 		grid-template-columns: 1fr;
 		justify-items: center;
-	}
-
-	.card {
-		max-width: 350px;
-	}
-
-	.show-search-options {
-		display: inline-block;
-	}
-
-	.refine-search {
-		display: none;
-	}
-
-	.cart-grid {
-		grid-template-columns: 1fr;
-	}
-
-	.cards-display-cart {
-		grid-template-columns: 1fr;
-		justify-items: center;
-	}
-
-	.header-search-bar {
-		width: auto;
 	}
 
 	.out-of-stock {
