+++ src/components/NavBar/NavBar.css
@@ -1,0 +1,69 @@
+.header {
+	padding: 0rem 1rem;
+	flex-wrap: wrap;
+}
+
+.header > h1 {
+	font-family: 'Roboto', sans-seridf;
+}
+
+.header-container {
+	position: relative;
+	box-shadow: none;
+}
+
+.header-search-bar {
+	padding: 0.5rem 1rem;
+	margin: 0.5rem auto;
+	min-width: 300px;
+	display: flex;
+}
+
+.header-search-bar > input {
+	outline: none;
+	border: none;
+	margin-right: 0.5rem;
+	width: 100%;
+}
+
+.header-search-bar > button {
+	outline: none;
+	border: none;
+	background: none;
+	font-size: 1rem;
+}
+
+.header-search-bar > button:hover {
+	color: var(--primary);
+}
+
+.links-bar {
+	display: flex;
+	padding: 1.2rem 0rem;
+	position: -webkit-sticky;
+	position: sticky;
+	top: 0;
+	background-color: var(--primary);
+	max-width: var(--max-width);
+	margin: 0 auto;
+	z-index: 3;
+}
+
+.links-bar > button {
+	background: none;
+	border: none;
+	padding: 0rem 1rem;
+	outline: none;
+	position: relative;
+	font-size: 1rem;
+}
+
+.links-bar > button:hover {
+	font-weight: bolder;
+}
+
+@media only screen and (max-width: 500px) {
+	.header-search-bar {
+		width: auto;
+	}
+}
