+++ src/pages/Cart/CartItemCard.js
@@ -1,49 +1,9 @@
-import { useCart } from '../contexts/CartProvider';
-
-export function Cart() {
-	const {
-		state: { cartItems, wishListItems },
-		dispatch,
-		dispatchWrapper,
-	} = useCart();
-
-	return (
-		<div>
-			<h3>Cart</h3>
-
-			<div className="cart-grid">
-				{cartItems.length === 0 ? (
-					<CartEmpty />
-				) : (
-					<div className="cards-display-cart">
-						{cartItems.map((item) => (
-							<ItemCard
-								key={item.id}
-								item={item}
-								dispatch={dispatch}
-								dispatchWrapper={dispatchWrapper}
-								wishListItems={wishListItems}
-							/>
-						))}
-					</div>
-				)}
-
-				<CartTotal cartItems={cartItems} />
-			</div>
-		</div>
-	);
-}
-
-function CartEmpty() {
-	return (
-		<div class="page-empty">
-			<h3>Cart Is Empty</h3>
-			<i class="fas fa-shopping-cart fa-5x"></i>
-		</div>
-	);
-}
-
-function ItemCard({ item, dispatch, dispatchWrapper, wishListItems }) {
+export function CartItemCard({
+	item,
+	dispatch,
+	dispatchWrapper,
+	wishListItems,
+}) {
 	function handleLikeButtonClick(item) {
 		if (
 			//if already in wishlist
@@ -176,54 +136,6 @@ function ItemCard({ item, dispatch, dispatchWrapper, wishListItems }) {
 					Remove
 				</button>
 			</div>
-		</div>
-	);
-}
-
-function CartTotal({ cartItems }) {
-	function calculateTotalPrice() {
-		const inStockItems = cartItems.filter((item) => item.inStock === true);
-
-		const price = inStockItems.reduce(
-			(acc, curr) => acc + parseInt(curr.price) * curr.quantity,
-			0
-		);
-
-		return price;
-	}
-	return (
-		<div className="cart-total border-shadow padding1">
-			<h5>Cart Total</h5>
-			<hr style={{ marginBottom: '1rem' }} />
-			<div className="flex-space-between margin-bottom-1">
-				<span>
-					Price (
-					{cartItems.filter((item) => item.inStock === true).length}{' '}
-					items)
-				</span>
-				<span>₹ {calculateTotalPrice()}</span>
-			</div>
-			<div className="flex-space-between margin-bottom-1">
-				<span>Discount</span>
-				{calculateTotalPrice() > 0 ? (
-					<span className="primary-text-color">- ₹40</span>
-				) : (
-					<span className="primary-text-color">- ₹0</span>
-				)}
-			</div>
-			<div className="flex-space-between margin-bottom-1">
-				<span>Delivery Charges</span>
-				<span className="primary-text-color">FREE</span>
-			</div>
-			<hr style={{ marginBottom: '1rem' }} className="hr-dashed" />
-			<div className="flex-space-between margin-bottom-1">
-				<span className="bold-font-weight">TOTAL</span>
-				<span className="bold-font-weight">
-					₹
-					{calculateTotalPrice() > 0 ? calculateTotalPrice() - 40 : 0}
-				</span>
-			</div>
-			<button className="btn btn-primary">CHECKOUT</button>
 		</div>
 	);
 }
